Split Azure DevOps test results client
